recommendation-system/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
scikit-learn
numpy
